Refactor Axis protocol and add Axis using macro so that axes can be accessed too (#19)
