+++ lib/plox/color_axis.ex
@@ -4,6 +4,8 @@ defmodule Plox.ColorAxis do
   with, so it should be documented
   """
 
+  use Plox.Axis
+
   alias Plox.ColorScale
 
   defstruct [:scale]
@@ -14,7 +16,7 @@ defmodule Plox.ColorAxis do
 
   # def values(%__MODULE__{scale: scale}, opts \\ %{}), do: Scale.values(scale, opts)
 
-  defimpl Plox.Axis do
+  defimpl Plox.Axis.Protocol do
     def to_graph(%{scale: scale}, value) do
       ColorScale.convert_to_color(scale, value)
     end
