+++ lib/plox/linear_axis.ex
@@ -4,6 +4,8 @@ defmodule Plox.LinearAxis do
   with, so it should be documented
   """
 
+  use Plox.Axis
+
   alias Plox.Scale
 
   defstruct [:scale, :min, :max]
@@ -18,7 +20,7 @@ defmodule Plox.LinearAxis do
 
   def values(%__MODULE__{scale: scale}, opts \\ %{}), do: Scale.values(scale, opts)
 
-  defimpl Plox.Axis do
+  defimpl Plox.Axis.Protocol do
     def to_graph(%{scale: scale, min: min, max: max}, value) do
       Scale.convert_to_range(scale, value, min..max)
     end
